fix : [blogController]

+++ src/main/java/icu/shishc/service/serviceImpl/BlogServiceImpl.java
@@ -1,11 +1,13 @@
 package icu.shishc.service.serviceImpl;
 
+import icu.shishc.Exception.CustomException;
 import icu.shishc.entity.Blog;
 import icu.shishc.enumeration.BlogStatus;
 import icu.shishc.mapper.BlogMapper;
 import icu.shishc.service.BlogService;
-import icu.shishc.util.MyUtils;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,6 +18,7 @@ import java.util.List;
  * @Auther:ShiShc
  */
 
+@Slf4j
 @Service
 public class BlogServiceImpl implements BlogService {
 
@@ -24,34 +27,49 @@ public class BlogServiceImpl implements BlogService {
 
 
     @Override
-    public Integer insert(Blog blog) {
-        Blog blogTemp = blogMapper.getBlogByTitle(blog.getTitle());
-        if(blogTemp != null) {
-            MyUtils.print("Can't create blogs with the same title!");
-            return -1;
+    public Blog insert(Blog blog) throws CustomException {
+        Blog blog1 = blogMapper.getBlogByTitle(blog.getTitle());
+        if(blog1 != null) {
+            log.warn("【Service】BlogService::insert:the blog has exist! blogTitle = {}", blog.getTitle());
+            throw new CustomException(HttpStatus.BAD_REQUEST, "Blog Has Exist!");
         }
         blogMapper.insert(blog.getUsername(), blog.getTitle(), blog.getContent(), blog.getStatus().getKey());
-        MyUtils.print("Blog created successfully");
+        Blog blog2 = blogMapper.getBlogByTitle(blog.getTitle());
+        log.info("【Service】BlogService::insert:add blog successfully! bID = {}", blog2.getBlogId());
+        return blog2;
+    }
+
+
+    @Override
+    public Integer delete(Long id) throws CustomException {
+        if(id <= 0) {
+            log.warn("【Service】BlogService::delete:Illegal param, bid = {}", id);
+            throw new CustomException(HttpStatus.BAD_REQUEST, "bad param");
+        }
+        blogMapper.delete(id);
+        log.info("【Service】BlogService::delete: delete blog, bid = {}", id);
         return 1;
     }
 
 
     @Override
-    public Integer delete(Long id) {
-        Integer flag = blogMapper.delete(id);
-        return 1;
-    }
-
-
-    @Override
-    public Integer update(Blog blog) {
-        blogMapper.update(blog.getTitle(), blog.getContent(), blog.getStatus().getKey(), blog.getBlogId());
-        return 1;
+    public Blog update(Blog blog) throws CustomException {
+        Long id = blog.getBlogId();
+        Blog blog1 = blogMapper.getBlogByBID(id);
+        if(null == blog1) {
+            log.warn("【Service】BlogService::update: the blog doesn't exist! bid = {}", id);
+            throw new CustomException(HttpStatus.BAD_REQUEST, "blog doesn't exist!");
+        }
+        blogMapper.update(blog.getTitle(), blog.getContent(), blog.getStatus().getKey(), id);
+        log.info("【Service】BlogService::update: update successfully! bID = {}", id);
+        Blog blog2 = blogMapper.getBlogByBID(id);
+        return blog2;
     }
 
 
     @Override
     public List<Blog> getAllBlog() {
+        log.info("【Service】BlogService::getAllBlog");
         List<Blog> blogList = blogMapper.getAllBlog();
         return blogList;
     }
@@ -60,35 +78,40 @@ public class BlogServiceImpl implements BlogService {
     @Override
     public Blog getBlogByTitle(String title) {
         Blog blog = blogMapper.getBlogByTitle(title);
-        MyUtils.print("Successfully found the blog by title");
+        log.info("【Service】BlogService::getByTitle: bid = {}", blog == null ? 0 : blog.getBlogId());
         return blog;
     }
 
 
     @Override
-    public Blog getBlogByBID(Long bid) {
+    public Blog getBlogByBID(Long bid) throws CustomException{
+        if(bid <= 0) {
+            log.warn("【Service】BlogService::getByID: bad bid = {}", bid);
+            throw new CustomException(HttpStatus.BAD_REQUEST, "bad bid");
+        }
         Blog blog = blogMapper.getBlogByBID(bid);
+        log.info("【Service】BlogService:getByID: bid = {}", bid);
         return blog;
     }
 
 
     @Override
     public List<Blog> getBlogByStatus(BlogStatus blogStatus) {
-        List<Blog> list = blogMapper.getBlogByStatus(blogStatus);
-        return list;
+        log.info("【Service】BlogService::getByStatus");
+        return blogMapper.getBlogByStatus(blogStatus.getKey());
     }
 
 
     @Override
     public Integer getAllLike() {
-        Integer allLike = blogMapper.getAllLike();
-        return allLike;
+        log.info("【Service】BlogService::getAllLike");
+        return blogMapper.getAllLike();
     }
 
 
     @Override
     public Integer getAllReadNum() {
-        Integer allReadNum = blogMapper.getAllReadNum();
-        return allReadNum;
+        log.info("【Service】BlogService::getAllReadNum");
+        return blogMapper.getAllReadNum();
     }
 }
